crud project

+++ read.php
@@ -1,11 +1,16 @@
 <?php
+session_start();
+if (isset($_SESSION['login'])) {
+} else {
+    header("location:login.php");
+}
+//* verify user login
+
 $connection = mysqli_connect("localhost", "root", "", "backendphp");
 $qurey = mysqli_query($connection, "SELECT * FROM `products`");
 $result = mysqli_fetch_all($qurey, MYSQLI_ASSOC);
 $id = $_POST['id'] ?? '';
 ?>
-
-
 
 <!DOCTYPE html>
 <html lang="en">
@@ -22,14 +27,29 @@ $id = $_POST['id'] ?? '';
 </head>
 
 <body>
+    <nav class="navbar bg-dark border-bottom border-body" data-bs-theme="dark" style="border-bottom: 0px solid black !important; ">
+        <div>
+            <a href="login.php" style="font-weight: bold;font-size:35px;color:white;text-decoration: none;">
+                C R U D
+            </a>
+        </div>
+        <div class="btns">
+            <a href="logout.php"><button type="button" class="btn btn-danger">Logout</button></a>
+            <?php if ($_SESSION['login']['admin']): ?>
+                <a href="add.php"><button type="button" class="btn btn-primary" id="add-button">Add New Product</button></a>
+            <?php endif; ?>
+        </div>
+    </nav>
     <table class="table" id="product-table">
         <thead class="thead-dark">
             <tr>
                 <th scope="col">ID</th>
                 <th scope="col">Product Title</th>
                 <th scope="col">Product Price</th>
-                <th scope="col">Update</th>
-                <th scope="col">Delete</th>
+                <?php if ($_SESSION['login']['admin']): ?>
+                    <th scope="col">Update</th>
+                    <th scope="col">Delete</th>
+                <?php endif; ?>
             </tr>
         </thead>
         <tbody>
@@ -38,14 +58,15 @@ $id = $_POST['id'] ?? '';
                     <th><?= $vlaue['id'] ?></th>
                     <td><?= $vlaue['title'] ?></td>
                     <td><?= $vlaue['price'] ?></td>
-                    <td><a href="update.php?id=<?= $vlaue['id'] ?>&title=<?= urlencode($vlaue['title']) ?>&price=<?= $vlaue['price'] ?>"><button type="button" class="btn btn-warning">Update</button></a></td>
-                    <td><a href="delete.php?id=<?= $vlaue['id'] ?>"><button type="button" class="btn btn-danger">Delete</button></a></td>
-
+                    <?php if ($_SESSION['login']['admin']): ?>
+                        <td><a href="update.php?id=<?= $vlaue['id'] ?>&title=<?= urlencode($vlaue['title']) ?>&price=<?= $vlaue['price'] ?>"><button type="button" class="btn btn-warning" style="color:white;">Update</button></a></td>
+                        <td><a href="delete.php?id=<?= $vlaue['id'] ?>"><button type="button" class="btn btn-danger">Delete</button></a></td>
+                    <?php endif; ?>
                 </tr>
             <?php endforeach; ?>
         </tbody>
     </table>
-    <a href="add.php"><button type="button" class="btn btn-primary" id="add-button">Add New Product</button></a>
+
 </body>
 
 </html>
